+++ app/means_test/views.py
@@ -2,7 +2,8 @@ from flask.views import View, MethodView
 from flask import render_template, url_for, redirect, session
 
 from app.means_test.api import update_means_test
-from app.means_test.forms import BenefitsForm, AboutYouForm
+from app.means_test.forms.about_you import AboutYouForm
+from app.means_test.forms.benefits import BenefitsForm
 
 
 class MeansTest(View):
